[docs] components 분리
인증예시등록 & 캘린더 & 대표이미지선택 컴포넌트 분리
